whole-chain envelope v6: the count lane no longer aliases — one edited integer no longer inflates a history
MEASURED OPEN (c80ee7c92, 0a1854df0): editing one `u64` in a `WholeChainProofBytes`
envelope — no key, no proving, no witness — made a genuine 2-turn history verify and
report `AttestedHistory.num_turns = 6,308,233,219`. 12/12 aliases admitted across the
envelope, blob and in-memory seams; the control `n + 1` refused, so it was attributable
to modular aliasing and not a dead tooth.

Cause: the count rides as ONE BabyBear lane. Both comparison sites build
`BabyBear::new(num_turns as u32)` — `as u32` truncates mod 2^32, `BabyBear::new` reduces
mod p — and every `num_turns < BABY_BEAR_MODULUS` guard in the file sat on a `turns.len()`
PROVER site (:2593, :3021, :3432, :5249). None on a verify path. The AIR-level version of
this same defect was closed earlier the same day by constraint 14; the wire never got it.

THE FIX, structural — three places, one of which the original audit missed:

1. `WholeChainProofBytes.num_turns` is now a `u32`, not a `u64`. The `2^32` family is dead
   BY TYPE: not representable on the wire at all.
2. `WholeChainProofBytes::from_postcard` refuses `num_turns >= BABY_BEAR_MODULUS`. This is
   the ONE decode gate every wire consumer passes, and the guard sits on the `u32` and not
   on a `usize` — under wasm32 a `usize` is 32 bits, so `env.num_turns as usize` would
   already have truncated before a `usize`-typed guard could see it.
3. `verify_turn_chain_recursive_from_parts_with_board_window` refuses it too, covering the
   four entries that funnel there — including `verify_turn_chain_recursive_from_blobs`, the
   pg-dregg seam that takes a bare `usize` and passes no decode gate.
4. `verify_wide_turn_chain_recursive` does NOT funnel; it builds its own
   `BabyBear::new(proof.num_turns as u32)` and gets its own branch. The first read of this
   fix covered only the funnelling entries and would have left this one aliasing.

`WholeChainProofBytes::from_proof` now panics rather than narrowing a count the wire cannot
carry faithfully. Every fold entry already refuses such a chain, so it fires only on a
hand-mutated `WholeChainProof` — where a truncating `as u32` is exactly the forgery.

WHAT BREAKS (the flag day):

* `WHOLE_CHAIN_PROOF_ENVELOPE_V1` 5 -> 6. Every v5 artifact now REFUSES TO LOAD
  ("unsupported envelope version 5 (this build reads v6)") rather than being reinterpreted.
* RE-EMIT: the envelope only. `WholeChainProofBytes::from_proof(&proof).to_postcard()` on
  the same fold. NO VK rotation, NO re-proving, NO descriptor re-emit, NO re-genesis — the
  root proof, the binding proof and the segment exposure are untouched.
* Anything holding stored `.to_postcard()` bytes must re-emit them: the wasm
  `ExternalHistoryEnvelope.proof_bytes_b64` payloads, pg-dregg's stored transports, any
  fixture blob on disk.
* `GnarkWitnessExportError::NumTurnsNotCanonical.value` narrowed u64 -> u32. Its
  "above u32::MAX" arm is now dead by type; the `>= p` arm stays live for hand-built
  envelopes. The 25-lane gnark public-input contract is UNCHANGED.
* `EthPublicInputsV2::new` still takes a `u64` (non-envelope callers hand it one), but the
  envelope seam `from_whole_chain_bytes` can no longer reach its "does not fit u32" arm.

BOTH PROBES INVERTED, DELIBERATELY. Their assertions were not edited to match the new
behaviour; each was rewritten to state the refusal the fix owes, so a regression fails
loudly instead of passing quietly. `num_turns_alias_probe.rs` gains three FAST legs that
need no fold: the decode gate refuses `>= p` and still admits `p - 1` (a bound, not a
blanket); a hand-built `u64`-shaped envelope carrying 6,308,233,219 does not decode at the
BYTES ("dead by type" measured about the wire, not about Rust); and a v5 artifact refuses
to load. The slow legs keep the honest baseline ADMITTED and require the control `n + 1` to
still be refused by the SEGMENT TOOTH rather than by the new bound — so the measurement
stays attributable to where it always was.

+++ bridge/src/ethereum.rs
@@ -706,7 +706,11 @@ impl EthPublicInputsV2 {
         Self::new(
             src.genesis_root,
             src.final_root,
-            src.num_turns,
+            // Envelope v6 narrowed this field `u64` -> `u32` (the count-lane alias close), so the
+            // "does not fit u32" arm of `new` is now unreachable FROM THIS SEAM — dead by type,
+            // not by check. `new` keeps taking a `u64` because callers that are not the envelope
+            // (raw settlement JSON, the interchain adapter) still hand it one.
+            u64::from(src.num_turns),
             src.chain_digest,
         )
     }
